+++ src/Node.cpp
@@ -1,13 +1,13 @@
 #include "Node.h"
 
-Meatball::Node::Node(int x, int y, int z, int width, int height, bool visible)
+Meatball::Node::Node(float x, float y, float z, float width, float height, bool visible)
 	: x(x), y(y), z(z), width(width), height(height), visible(visible) {}
 
 void Meatball::Node::update() {}
 void Meatball::Node::draw() {}
 
 bool Meatball::Node::checkCollision(Rectangle& rectangle) {
-	Rectangle thisRect = { (float)x, (float)y, (float)width, (float)height };
+	Rectangle thisRect = { x, y, width, height };
 	return CheckCollisionRecs(thisRect, rectangle);
 }
