+++ src/Panel.cpp
@@ -1,6 +1,6 @@
 #include "Panel.h"
 
-Meatball::Interface::Panel::Panel(int x, int y, int width, int height, Color color, bool visible)
+Meatball::Interface::Panel::Panel(float x, float y, float width, float height, Color color, bool visible)
 	: NodeUI(x, y, width, height, visible), color(color) {}
 
 int Meatball::Interface::Panel::getTypes() {
