+++ src/Panel.h
@@ -6,7 +6,7 @@ namespace Meatball {
 	namespace Interface {
 		class Panel : public NodeUI {
 		public:
-			Panel(int x, int y, int width, int height, Color color, bool visible = true);
+			Panel(float x, float y, float width, float height, Color color, bool visible = true);
 			
 			virtual void draw();
